readme2
Tries are used to implement lookups in maps keyed on character data quickly.

The canonical example of using a trie in a search engine is search box type ahead. As the user types in the search box, you can look up the characters typed so far in the trie and store the most likely searches based on the characters typed so far at each node in the trie.

You can do this whether your search engine runs on your local machine or not, but doing this requires machine learning and data to try to figure out the most likely searches.

The search itself could be implemented by storing all the possible words in all the pages you crawled in a trie if you have enough space for that, but more likely a simple, single machine implementation of a search engine will just search all the pages it has crawled and you would only worry about building indexes to speed up the queries if you want to make it fast.